Acabada semana

+++ LaNubeApi/src/main/java/com/example/demo/repository/AlumnoRepo.java
@@ -3,8 +3,7 @@ package com.example.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.example.demo.model.Alumno;
-import com.example.demo.model.Usuario;
 
-public interface AlumnoRepo extends JpaRepository<Alumno, Long> {
+public interface AlumnoRepo extends JpaRepository<Alumno, Integer> {
 
 }
